% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_bootstrap.R
\name{dataGen.case}
\alias{dataGen.case}
\alias{dataGen.NP}
\alias{dataGen.Par}
\alias{sampDist.Par}
\alias{sampDist.NP}
\alias{sampDist.NPcond}
\title{Bootstrap datasets}
\usage{
dataGen.case(saemixObject)

dataGen.NP(saemixObject, nsamp = 0, eta.sampc = NULL, conditional = FALSE)

dataGen.Par(saemixObject)
}
\arguments{
\item{saemixObject}{an object returned by the \code{\link{saemix}} function}

\item{nsamp}{number of samples from the conditional distribution (for method="conditional")}

\item{eta.sampc}{if available, samples from the conditional distribution (otherwise, they are obtained within the function)}

\item{conditional}{if TRUE, sample from the conditional distribution, if FALSE, sample within the
empirical Bayes estimates (EBE) as in the traditional non-parametric residual bootstrap}
}
\description{
These functions create bootstrapped datasets using the bootstrap methods described in \code{\link{saemix.bootstrap}}
}
