% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixSimData.R, R/SaemixData.R, R/SaemixRes.R, R/SaemixObject.R
\docType{methods}
\name{[,SaemixData-method}
\alias{[}
\alias{[,SaemixData-method}
\alias{[,SaemixSimData-method}
\alias{[,SaemixRepData-method}
\alias{[,SaemixModel-method}
\alias{[,SaemixRes-method}
\alias{[,SaemixObject-method}
\alias{[,SaemixData}
\alias{[,SaemixRepData}
\alias{[,SaemixSimData}
\alias{[,SaemixModel}
\alias{[,SaemixRes}
\alias{[,SaemixObject}
\alias{[,SaemixSimData-method}
\alias{[<-,SaemixSimData-method}
\alias{[<-,SaemixData-method}
\alias{[<-,SaemixRepData-method}
\alias{[<-,SaemixModel-method}
\alias{[<-,SaemixRes-method}
\alias{[<-,SaemixObject-method}
\alias{[<-,SaemixData}
\alias{[<-,SaemixSimData}
\alias{[<-,SaemixRes}
\alias{[<-,SaemixModel}
\alias{[<-,SaemixObject}
\alias{[<--methods}
\alias{[-methods}

\title{Get/set methods for SaemixData object}
\usage{
\S4method{[}{SaemixData}(x, i, j, drop)

\S4method{[}{SaemixData}(x, i, j) <- value

\S4method{[}{SaemixSimData}(x, i, j, drop)

\S4method{[}{SaemixSimData}(x, i, j) <- value

\S4method{[}{SaemixRes}(x, i, j, drop)

\S4method{[}{SaemixRes}(x, i, j) <- value

\S4method{[}{SaemixObject}(x, i, j, drop)

\S4method{[}{SaemixObject}(x, i, j) <- value
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s)}

\item{i, j, ...}{indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details}

\item{value}{typically an array-like R object of a similar class as x}
}
\description{
Access slots of a SaemixData using the object["slot"] format

replace names of SaemixData
}
\keyword{methods}
