% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaemixData.R, R/SaemixModel.R, R/aaa_generics.R
\docType{methods}
\name{print,SaemixData-method}
\alias{print,ANY-method}
\alias{print,SaemixData-method}
\alias{print,SaemixModel-method}
\alias{print,SaemixRes-method}
\alias{print,SaemixObject-method}
\alias{print-methods}
\alias{print.saemix}
\title{Methods for Function print}
\usage{
\S4method{print}{SaemixModel}(x, ...)
}
\arguments{
\item{x}{an object of class SaemixData, SaemixSimData or SaemixObject}

\item{\dots}{optional arguments for displaying the output. By default only the first 10 lines of data are printed (nlines=10).}
}

\description{
Prints a summary of an object
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ Default print function }

\item{list("signature(x = \"SaemixData\")")}{ Prints a summary of a
SaemixData object }

\item{list("signature(x = \"SaemixModel\")")}{ Prints a summary of a
SaemixModel object }

\item{list("signature(x = \"SaemixObject\")")}{ Prints a summary of the
results from a SAEMIX fit }

\item{list("signature(x = \"SaemixRes\")")}{ Not user-level } }
}
\keyword{methods}

