% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaemixModel.R
\docType{methods}
\name{[,SaemixModel,ANY,ANY-method}
\alias{[,SaemixModel,ANY,ANY-method}
\title{Get/set methods for SaemixModel object}
\usage{
\S4method{[}{SaemixModel,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{object}

\item{i}{element to be replaced}

\item{j}{element to replace with}

\item{drop}{whether to drop unused dimensions}
}
\description{
Access slots of a SaemixModel object using the object["slot"] format
}
\keyword{methods}

