% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaemixRes.R
\docType{methods}
\name{[,SaemixRes,ANY,ANY-method}
\alias{[,SaemixRes,ANY,ANY-method}
\title{Get/set methods for SaemixRes object}
\usage{
\S4method{[}{SaemixRes,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{object}

\item{i}{element to be replaced}

\item{j}{element to replace with}

\item{drop}{whether to drop unused dimensions}
}
\description{
Access slots of a SaemixRes object using the object["slot"] format
}
\keyword{methods}

