% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaemixData.R
\name{subset.SaemixData}
\alias{subset}
\alias{subset-methods}
\alias{subset.SaemixData}
\title{Data subsetting}
\arguments{
\item{x}{saemixData object}

\item{subset}{logical expression indicating elements or rows to keep: missing values are taken as false}

\item{...}{additional parameters (ignored)}
}
\value{
an object of class \code{"\linkS4class{SaemixData}"}
}
\description{
Return an SaemixData object containing the subset of data which meets conditions.
}
\examples{
# TODO
}
\keyword{methods}

