% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaemixModel.R, R/aaa_generics.R
\docType{methods}
\name{summary,SaemixModel-method}
\alias{summary,ANY-method}
\alias{summary,SaemixModel-method}
\alias{summary,SaemixObject-method}
\alias{summary-methods}
\title{Methods for Function summary}
\usage{
\S4method{summary}{SaemixModel}(object, print = TRUE, ...)
}
\arguments{
\item{object}{an object of class SaemixModel or SaemixObject}
\item{print}{if TRUE, shows the output with pretty formatting}
\item{...}{additional arguments usually passed on to summary functions such as na.rm=TRUE}
}
\value{
A list with named elements from the object, summarising estimated parameters, random effects, etc...
}
\description{
Methods for function \code{summary}
}
\section{Methods}{
 \describe{

\item{list("signature(x = \"ANY\")")}{ default summary function ?}

\item{list("signature(x = \"SaemixData\")")}{ summary of the data }

\item{list("signature(x = \"SaemixModel\")")}{ summary of the model }

\item{list("signature(x = \"SaemixObject\")")}{ summary of an SaemixObject}

}
}
\keyword{methods}

