% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SaemixData.R, R/SaemixModel.R, R/SaemixObject.R, R/SaemixRes.R, R/aaa_generics.R
\docType{methods}
\name{initialize,SaemixData-method}
\alias{initialize,SaemixData-method}
\alias{initialize,SaemixModel-method}
\alias{initialize,SaemixObject-method}
\alias{initialize,SaemixRepData-method}
\alias{initialize,SaemixRes-method}
\alias{initialize,SaemixSimData-method}
\alias{initialize-methods}
\title{Methods for Function initialize}
\description{
Constructor functions for Classes in the saemix package
}
\section{Methods}{
 \describe{

\item{list("signature(.Object = \"SaemixData\")")}{ create a SaemixData
object. Please use the \code{\link{saemixData}} function.}

\item{list("signature(.Object = \"SaemixModel\")")}{ create a SaemixModel
object Please use the \code{\link{saemixModel}} function.}

\item{list("signature(.Object = \"SaemixObject\")")}{ create a SaemixObject
object. This object is obtained after a successful call to
\code{\link{saemix}}}

\item{list("signature(.Object = \"SaemixRepData\")")}{ create a
SaemixRepData object}

\item{list("signature(.Object = \"SaemixRes\")")}{ create a SaemixRes
object}

\item{list("signature(.Object = \"SaemixSimData\")")}{ create a
SaemixSimData object} }
}
\keyword{methods}

